--- src/larceler/LarCelerStandalone.cc
@@ -1,68 +1,0 @@
-//------------------------------- -*- C++ -*- -------------------------------//
-// Copyright Celeritas contributors: see top-level COPYRIGHT file for details
-// SPDX-License-Identifier: (Apache-2.0 OR MIT)
-//---------------------------------------------------------------------------//
-//! \file larceler/LarCelerStandalone.cc
-//---------------------------------------------------------------------------//
-
-#include "LarCelerStandalone.hh"
-
-#include <memory>
-#include <lardataobj/Simulation/OpDetBacktrackerRecord.h>
-#include <lardataobj/Simulation/SimEnergyDeposit.h>
-
-#include "corecel/Assert.hh"
-
-#include "LarStandaloneRunner.hh"
-
-namespace celeritas
-{
-//---------------------------------------------------------------------------//
-/*!
- * Construct with fcl parameters.
- */
-LarCelerStandalone::LarCelerStandalone(Parameters const& config)
-    : runner_inp_{from_config(config())}
-{
-}
-
-//---------------------------------------------------------------------------//
-/*!
- * Start Celeritas at the beginning of the job.
- */
-void LarCelerStandalone::beginJob()
-{
-    CELER_EXPECT(!runner_);
-    runner_ = std::make_unique<LarStandaloneRunner>(
-        std::forward<LarStandaloneRunner::Input>(runner_inp_));
-}
-
-//---------------------------------------------------------------------------//
-/*!
- * Run Celeritas on a single event.
- */
-auto LarCelerStandalone::executeEvent(VecSED const& edeps) -> UPVecBTR
-{
-    CELER_EXPECT(runner_);
-    CELER_EXPECT(!edeps.empty());
-
-    using VecBTR = LarStandaloneRunner::VecBTR;
-
-    // Calculate detector responsors for the input steps
-    auto& run = *runner_;
-    VecBTR result = run(edeps);
-    return std::make_unique<VecBTR>(std::move(result));
-}
-
-//---------------------------------------------------------------------------//
-/*!
- * Free Celeritas memory at the end of the job.
- */
-void LarCelerStandalone::endJob()
-{
-    CELER_EXPECT(runner_);
-    runner_.reset();
-}
-
-//---------------------------------------------------------------------------//
-}  // namespace celeritas
